internal/httpclient: Import GitHub CLI's API "HTTP Client" utility funcs
This commit imports related functionality found within the GitHub's CLI API
package for creating a standard HTTP client[0][1].  The new package is
extracted and placed within the `internal/httpclient` directory.  The copy
is partially direct and makes necessary copyright attribution which is
facilitated by placing the contents of the project's LICENSE (MIT) as a
comment preamble along with SPX identifier.

Modifications to the library have been made, including:

 - Removing GitHub-centric User-Agent identifiers;
 - Removing defaults and authentication from client requests, specific
   to GitHub's API;
 - Removing unused methods related to GitHub's API.

[0]: https://github.com/cli/cli/tree/1eefda0/api/client.go
[1]: https://github.com/cli/cli/tree/1eefda0/api/cache.go

Signed-off-by: Alexander Jung <alex@unikraft.io>