internal/ghrepo: Import GitHub CLI representation of a GitHub repo
This commit directly imports a slightly modified package from GitHub's CLI
tool and places it within the `internal/ghrepo` directory[0].  The copy is
partially direct and makes necessary copyright attribution which is
facilitated by placing the contents of the project's LICENSE (MIT) as a
comment preamble along with SPX identifier.

The primary modification to the package has been removing the dependency to
GitHub CLI's internap `ghinstance` package which aids in distinguishing
between GitHub's public API and an Enterprise "self-hosted" API.  For now,
such a differentiation is not necessary.

[0]: https://github.com/cli/cli/tree/1eefda0/internal/ghrepo

Signed-off-by: Alexander Jung <alex@unikraft.io>