internal/ghrepo: Import GitHub CLI config manager
This commit directly imports GitHub's CLI tool's ability to manage
configuration for the application itself verbatim and places it within the
`internal/config` directory[0].  The copy is partially direct and makes
necessary copyright attribution which is facilitated by placing the
contents of the project's LICENSE (MIT) as a comment preamble along with
SPX identifier.  It also does not include tests or stubs.

[0]: https://github.com/cli/cli/tree/1eefda0/internal/config

Signed-off-by: Alexander Jung <alex@unikraft.io>