internal/cmdfactory: Introduce `cmdfactory`
The command factory is an internal utility package which provides a higher-
order level of abstraction to new commands, allowing the command to access
commonly exposed interfaces accross the repository.  For now, the
implementation is minimal and simply includes directives such as access to
common config, a HTTP client as well as knowing the working directory.

The motive behind this package is to be able to introduce `With` option
arguments to the factory during initialisation and to further extend the
command's reach to different subsystems.  For example, a new command can
instantiate itself with a command factory which has access to the plugin
manager, or the manifest manager, etc.  Some commands may need access to
these common subsystems, others may not.  The command factory can help
standardise access to such features.

Signed-off-by: Alexander Jung <alex@unikraft.io>