internal/cmdutil: Partial import of GitHub CLI `cmdutil` helper methods
This commit directly imports GitHub's CLI tool's `cmdutil`[0] helper methods
to be used along with a general utility function which wraps the spf13
cobra command package.  These functions are copied verbatim with minimal
modification to remove packages which are non-dependent (mainly the
GitHub's extension manager) and makes necessary copyright attribution which
is facilitated by placing the contents of the project's LICENSE (MIT) as a
comment preamble along with SPX identifier.

[0]: https://github.com/cli/cli/tree/1eefda0/pkg/cmdutil

Signed-off-by: Alexander Jung <alex@unikraft.io>