internal/cmd/alias/expand: Import GitHub CLI command alias expander
This commit directly imports GitHub's CLI tool's command for expanding
aliases within the CLI itself.  The copy is direct and places the package
within `internal/cmd/alias/expand`[0] and makes necessary copyright
attribution which is facilitated by placing the contents of the project's
LICENSE (MIT) as a comment preamble along with SPX identifier.

[0]: https://github.com/cli/cli/tree/1eefda0/pkg/cmd/alias/expand

Signed-off-by: Alexander Jung <alex@unikraft.io>