internal/cmd/version: Import GitHub CLI command for version check
This commit directly imports GitHub's CLI tool's command for checking the
program version of the tool via the command-line.  The copy is direct and
places the package within `internal/cmd/version`[0] and makes necessary
copyright attribution which is facilitated by placing the contents of the
project's LICENSE (MIT) as a comment preamble along with SPX identifier.

[0]: https://github.com/cli/cli/tree/1eefda0/pkg/cmd/version

Signed-off-by: Alexander Jung <alex@unikraft.io>