cmd/kraftkit: Introduce minimal `kraftkit` exectuable
This initial commit of the `kraftkit` command uses the repository's
internal methods to generate a guided and consistent program.  Using the
Make target `make kraftkit` will build this binary.  The program itself
does nothing at this point other than show a help menu, however, additional
modifications will be made in the future to be able to manage the whole
toolsuite provided by the repository "kraftkit" itself.

Signed-off-by: Alexander Jung <alex@unikraft.io>