cmd/ukpkg: Introduce minimal `ukpkg` exectuable
This initial commit of the `ukpkg` command uses the repository's
internal methods to generate a guided and consistent program.  Using the
Make target `make ukpkg` will build this binary.  The program itself
does nothing at this point other than show a help menu, however, additional
modifications will be made in the future to be able to manage the whole
packages and components for build and running Unikraft unikernels.

Signed-off-by: Alexander Jung <alex@unikraft.io>