pkg/plugins: Introduce package for declaring and managing plugins
Inspired by GitHub's CLI tool which has "Extensions" feature, this package
follows a similar design pattern however it utilises Golang's native
plugin mechanism.  This allows us to compile programs with KraftKit
functionality and invoke it with the ability to literally extend any part
of KraftKit (as long as relevant public exposure by a package allows
extending).

This commit introduces an "essentials" version of the plugin package, with
limitations including:

 * No parsing of remote repositories which could provide functionality;
 * Not parsing a "manifest" file with details of the plugin itself.

Signed-off-by: Alexander Jung <alex@unikraft.io>