schema: Import the compose specification's schema validator
This commit imports the schema validator from the compose specification's
schema validator[0] as kraft's previous implementation use a similar
approach to validating the provided YAML file.  This commit is verbatim
with a change of pointing to the "embed" schema file and makes necessary
copyright attribution which is facilitated by placing the contents of the
project's LICENSE (Apache-2.0) as a comment preamble along with SPDX
identifier.

[0]: https://github.com/compose-spec/compose-go/blob/48546e0/schema/schema.go

Signed-off-by: Alexander Jung <alex@unikraft.io>