schema/loader: Introduce `LoaderOptions` for aiding to load a spec
This commit imports a partial from the compose specification loader[0]
which is useful for managing options for the loader program.  This program
is intended to read and parse a YAML specification file.  This commit also
makes necessary copyright attribution which is facilitated by placing the
contents of the project's LICENSE (Apache-2.0) as a comment preamble along
with SPDX identifier.

[0]: https://github.com/compose-spec/compose-go/blob/32e1258/loader/loader.go#L46-L79=

Signed-off-by: Alexander Jung <alex@unikraft.io>