schema/loader: Introduce ability to load and parse YAML schema
This commit introduces essential functionality to parse a `kraft.yaml` file
based on the compose specification project[0].  Necessary functionality to
perform the validation, interpolation and parsing are brought from the
project whilst simultaneously making necessary copyright attribution which
is facilitated by placing the contents of the project's LICENSE
(Apache-2.0) as a comment preamble along with SPDX identifier.

The code to perform loading in the schema is adapted to the kraft
specification and thus makes necessary changes in the same commit in order
to preserve functionality.  Additional work is necessary to ensure
compatibility against previous versions as well as handling additional
attributes.  This commit represents minimal parsing of the specification.

[0]: https://github.com/compose-spec/compose-go/tree/32e1258

Signed-off-by: Alexander Jung <alex@unikraft.io>