pkg/unikraft/component: Introduce "generic" component
This generic is used as both to provide abstract interfaces to future
components which consist of a Unikraft unikernel (e.g. lib, the core, etc.)
as well as common metadata which would be shared and equal amongst these
components (e.g. the fact that each component has a name, a source origin,
etc.).

Signed-off-by: Alexander Jung <alex@unikraft.io>