pkg/pkgmanager: Introduce the abstract package manager
This commit introduces a new package within KraftKit for use of abstracting
different types of "package managers".  This functionality allows one to
have different methods by which we can find and retrieve unikernel
packages.

At the same time, a "umbrella" manager is introduced which is used to
iteravely call the same methods adopted by the `PackageManager` interface
via the global `RegisterPackageManager` method.

Signed-off-by: Alexander Jung <alex@unikraft.io>