pkg/unikraft: Migrate `MappingWithEquals` to `KConfig`
When this code was imported from the Docker Compose specificaiton, it was
intended to be used for environmental variables.  Similar to env vars is
the structure of Unikraft's KConfig configuration management system.  This
commit simply renames this functionality and places it within the already
defined `KConfig` type.  We can use this structure with the previous
migration of "`Environment`" to "`Configuration`" seen in previous commits.

Signed-off-by: Alexander Jung <alex@unikraft.io>