pkg/manifest: Introduce minimal "Manifest Manager" implementation
This commit introduces a new package manager known as "manifest manager"
which will be the default package manager within KraftKit.  In this commit,
the interfaces are realised but return an "not implemented" error message.
The intended implementation of this package manager is to deal with
tarballs of source directories of each component and arguably represents
the simpelst form of dealing with the code used to construct a Unikraft
unikernel.

Signed-off-by: Alexander Jung <alex@unikraft.io>