internal/cmdutil: Introduce `VarPF` and `StringVarP`
This commit introduces two command utility methods which build upon pflag's
internal flag organization system.  The main difference here is that we
return the `*pflag.Flag` rather than letting pflag internalize it.  We can
use the result of `VarPF` and `StringVarP` ourselves and inject it later on
based on a required desired state, for example with plugins which wish to
inject a new cmdline flag.  Since these methods require a pointer to a
variable, we can also track its evaluated value once the command is
executed.

Signed-off-by: Alexander Jung <alex@unikraft.io>