internal/{config,cmdfactory,httpclient}: Refactor configuration management
This commit provides "a new and improved" internal configuration
managemente system that is type-based.  The previous implementation of
configuration management was provided in 8b3b6bf and originates from
GitHub's CLI tool ("gh").  The main limitation of their implementation was
the constant need to cast the resulting type or provide the config getter
interface (as now removed from `httpclient`).  The interface was clunky and
always required error checking for access to the specific configuration.

In this commit and implementation, we introduce a `ConfigManager` for
reading and writing from various "feeds" of configuation as well as a
well-defined `Config` structure.  When we are looking up well-defined
configuration values, we need not check it merely exists as we will always
have some return (whether that's the default or the value passed in by a
feeder).  In the future, plugins may be able to introduce abstract
interfaces into `Config` and in this case only will we need to run a check
but if the plugin is compiled it will know the configuration it is looking
for by type-casting it ahead of time.

This commit and implementation has simplified access to the config in the
`cmdfactory`.

Signed-off-by: Alexander Jung <alex@unikraft.io>