pkg/pkgmanager: Introduce constructor and arch/plat pull options for PM
This commit introduces an exported `NewPackageManagerOptions` method which
can be used to customize the instantiation of the package manager.  In
addition, two options are introduced: `WithPullPlatform` and
`WithPullArchitecture` which pre-defined the platform and architeture to
pull.  The idea here is that we could theoretically pull a wide selection
of candidates from where the package manager retrieves its sources and
these options act as filters.

Signed-off-by: Alexander Jung <alex@unikraft.io>