pkg/log,internal/logger: Internalize logger package
The logger package need not be exported by the project.  This commit
migrates the package into the internal/ folder to keep it within the
project.  At the same time, to prevent symbol conflicts it has been renamed
from `log` to `logger`.  All structs using the logger should have a field
named `log` as a standard across KraftKit to ensure consistency.

Signed-off-by: Alexander Jung <alex@unikraft.io>