internal/config: Introduce compulsory feeder `Write` method
This commit adds a new method `Write` to the feeder interface as well as
includes an implementation for the current feeder based on YAML.  The
implementation recursively parses the input interface and replaces the
current file's values with what it is passed to it.

For now, this implementation does not maintain formatting or comments and
this is left for future work to formalize in the scenario where user's have
customized the configuration file.

Signed-off-by: Alexander Jung <alex@unikraft.io>