internal/logger,pkg/log: Introduce an abstract interface for the logger
Golang currently does not export a generic logging interface, meaning that
potential users who wish to use KraftKit's exposed modules inside of pkg/
are completely reliant on the internal logging system.  This may not suite
the context of the importing program.  In this commit, we introduce an
abstract and exported `Logger` interface within `pkg/log` such that we can
pass in other loggers and logging systems which adopt these methods.

Signed-off-by: Alexander Jung <alex@unikraft.io>