internal/{cmdfactory,config},pkg/plugins: Preemptively determine plugin dir
During command-line initializaiton (via the cmdfactory), determine the path
of the plugin directory by looking at the configuration.  If it is set, we
pass this rather than letting the `PluginManager` second guess where it is.
At the same time, we cache the `PluginManager` with this information and
clean up the configuration paths via `omitempty` for paths.

Signed-off-by: Alexander Jung <alex@unikraft.io>