pkg/manifest: Introduce initial structure of manifests
This commit introduces the initial structure of the manifest to be used as
the primary way of representing components within the Unikraft ecosystem.

The manifest is represented as one of two types of YAML file: an index
(which contains a list of manifests) and a single manifest file.  The
manifest itself contains the metadata of the component, including name,
type, description and known channels and versions.  A channel is a way of
representing a constant name (e.g. "staging" or "stable") which can be
continiously referenced.  Upon a new version, the channel can be updated to
select this version.  This way a user who wishes to subscribe a component
to a particular channel need not worry about setting the specific version.
This is similar to the concept of `latest` with Docker images, other than
version is never lost (since it remains in the list of known versions).

Signed-off-by: Alexander Jung <alex@unikraft.io>