config,internal/config: Expose `Config` as global package
This commit exposes the internal `Config` structure as a global package.
It allows accessing and setting the configuration with other packages which
may need access to configuration options and are available as public
packages themselves.  In addition, `Config` is exposed at the top-level
directory and not within `pkg/`.  This is the first commit of a series
which will migrate all existing `pkg/` packages out of this directory and
into the top-level directory.

Signed-off-by: Alexander Jung <alex@unikraft.io>