pkg/manifest: Introduce `Provider` interface with `NewProvider`
The `Provider` interface represents a mechanism within the `manifest`
package which will ultimatelly allow the `ManifestManager` to be able to
retrieve information resulting in a populated `Manifest` structure.  Since
the `ManifestManager` can accept as its input source either a Git
repository (in parity to the way in which the current version of the
`kraft` operates) as well as the new YAML-defined meta files whose
structure is defined in `Manifest` and `ManifestIndex`, the `NewProvider`
method (and subsequent implementing providers) will be able to accept as
input a `path` which can be tested against different providers.  A provider
which understands the input `path` can then call the implementing
interface's `Manifests()` method which returns a list of populated
`Manifest` structures.

Signed-off-by: Alexander Jung <alex@unikraft.io>