config: Add a map of `AuthConfig`s to the general `Config` struct
When authenticating against a service, `AuthConfig` will come in handy as a
way to retrieve the most vital pieces of information which is generalized
for most services.  A map of this struct is provided within the general
`Config` struct such that the key represents the domain or service name to
authenticate against.  This struct can be matched and passed to any
relevant service.

Signed-off-by: Alexander Jung <alex@unikraft.io>