pkg/manifest: Introduce `FindManifestsFromSource`
This method (and the internal method `findManifestsFromSource`) is a
recursive method which uses the introduced `NewProvider` method to validate
a given source.  Matching against a provider ensures the validity of the
source but also allows us to read it and return all the Manifests it may
represent.  For example, a ManifestIndex returning many, or a single
Manifest file, or even a Git repository representing only 1 Manifest.

Signed-off-by: Alexander Jung <alex@unikraft.io>