pkg/unikraft/component: Introduce `ParseComponentConfig` to parse iface
This commit introduces a new method `ParseComponentConfig` which is used
to convert an abstract `interface{}` into a `ComponentConfig`.  This
method may not necessarily be strictly necessary for now and exists to
satisfy commonality with equal methods for other components which are
are used in the schema's transformation process.  The mapstructure
library is able to handle attribute matching so the switch here is
redundant.  We keep this for now so that at least a string-based prop
can be cast to a `ComponentConfig`'s `Version`.

Signed-off-by: Alexander Jung <alex@unikraft.io>