schema: Fix parsing of target schema
This commit fixes how the target schema was parsed.  Previously, the
target itself was initially parsed.  With the migration of the target as
a "child" of a `ComponentConfig`, we must the therefore pass this first
so we can propagate attributes.

In addition, we remove the use of a `switch true` statement as this made
no sense in retrospect: we want to check all conditions of missing
attributes and populate accordingly.

Signed-off-by: Alexander Jung <alex@unikraft.io>