pkg/pkg: Introduce `ContextKey` for packages
The `PackageOptions` struct has a `context.Context` attribute which can
be used to pass package-specific implementation values to implementing
package managers.  This is different from metadata which is just
information.  With Context's `Value` method, we are able grab the whole
interface (which, if instantiated, can contain method pointers) of the
implementing package.  To make this process streamlined, we export a
`ContextKey` type which can be uniquely set for the implementing
package, such that we can lookup any additional information the
implementing pacakge provides.

Signed-off-by: Alexander Jung <alex@unikraft.io>