pkg/{manifest,pkgmanager}: Introduce `AddSource` method to interface
This commit introduces a new interface method which is implemented by
the `ManfiestManager` which allows one to source a package to the
manager.  This "feeds" the package manager with information about the
location of a or group of packages.  In the case of the
`ManifestManager` this is either a repository or a YAML manifest file
which can list many other manifests.

Signed-off-by: Alexander Jung <alex@unikraft.io>