pkg/{manifest,pkgmanager}: Register packages globally using `ContextKey`
Now that we have an established type to uniquely identify internally a
particular package manager, we can use this representation rather than a
string representation.  The use of `ContextKey` is also used to uniquely
identify contextual information passed into a package by the particular
package manager so it makes sense to keep this identification
consistent.

Signed-off-by: Alexander Jung <alex@unikraft.io>