pkg/manifest: Introduce `NewPackageWithVersion` method
This commit introduces a new method which, given a specific input
version string (which could either be a channel name or indeed a version
string), traverses through the provided `Manifest` and reduces it to
this input version only.  This information is packed into the
`PackageOptions` via the `ManifestContext` `ContextKey`.  Since this
becomes the only "version" context of the package and within the
`Manifest`, it can be easily read later when generating the actual
package.

Signed-off-by: Alexander Jung <alex@unikraft.io>