internal/cmdutil: Temporarily introduce static CLI options for logging
This is a temporary introduction of global CLI arguments to manipulate
the logger's configuration.  These options should, as described in an
inline TODO, be automatically generated (and grouped) by parsing the
global configuration.  This means that introducing a new configuration
option automatically adds new global CLI options and so the effort is a
1-op.  For now, this solution suffices until this logic can be
implemented more concretely.

Signed-off-by: Alexander Jung <alex@unikraft.io>