internal/cmdutil: Introduce independent `BoolVarP`
This commit introduces `BoolVarP` which returns the `*pflag.Flag` struct
such that it can be injected as argument dynamically.  The intt ernal
implementation references an internal structure and registers this
there which does not allow for attaching it to the `cobra.Command` of
choice, instead the one that has been instantiated internally.

Signed-off-by: Alexander Jung <alex@unikraft.io>