ui/paraprocess: Introduce generic "Parallel Process Progressbar" UI
This commit introduces an initial UI component for handling parallel
processes which expose a progress meter.  The UI can be used, for
example, to express downloads of components.  It is configurable through
a number of exposed options, including the ability to completely disable
whether processes occur in parallel; whether to render or not render via
"fancy UI" which may be used in a CI/CD context; the function used to
update the progress; and, access to a instantiated logger to print
messages to.

Signed-off-by: Alexander Jung <alex@unikraft.io>