ui/processtree: Introduce "process tree" UI component
The "Process Tree" UI component is used to manage processes which are
dependent on one-another.  The instantiation of processes occurs
depth-first, meaning child processes are executed before their parent
process.  Handling of success of child processes propagates upwards,
meaning that a parent process cannot start if any of the child processes
have failed.  The package exposes a number of options, including
necessary access to a logger (such that inline logging can occur);
whether to processes are executed in parallel; and whether to render the
output in a manner friendly for linear logging execution (e.g. within
the context of a CI/CD).

Signed-off-by: Alexander Jung <alex@unikraft.io>