config: Remove unused methods
This commit removes two methods `parseConfigFile` and `parseConfigData`
which were necessary for the parsing of the confiuration files
originally imported from GitHub's CLI tool.  Since this import, however,
the method in which the configuration files are retrieved and parsed
have changed into a custom "feeder" system and so these methods are no
longer necessary.

Signed-off-by: Alexander Jung <alex@unikraft.io>