cmd/ukpkg: Introduce the root `ukpkg` command
This is the initial commit of the `ukpkg` command.  With `ukpkg` you
are able to turn output artifacts from `ukbuild` into a
distributable archive ready for deployment.  At the same time,
`ukpkg` allows you to manage these archives: pulling, pushing, or
adding them to a project.

The default behaviour of `ukpkg` is to package a project.  Given no
arguments, you will be guided through interactive mode.

Github-Closes: #29
Signed-off-by: Alexander Jung <alex@unikraft.io>