pack,manifest: Introduce mandatory `Package` `Pack` method
The `Pack` method of an implementing `Package` interface is intended to
generate a distributable package given the implementation.  So far this
has not been implemented for the manifest package but acts as a
temporary placeholder.  In the future, additional options may be
necessary and optional for the creation of a package.

Signed-off-by: Alexander Jung <alex@unikraft.io>