chore: Output result from YTT to different file
This commit changes the path of the output from YTT and the generated
goreleaser file from `.goreleaser-staging.yaml` to
`goreleaser-staging.yaml`.  The file name is arbitrary, so we simply
remove the fact that it is hidden.  However, the change is important as
the GitHub Actions pipeline does not seem to recognise the resulting
difference and yields a non-recoverable error otherwise.  This
modifcation ensures GoReleaser can indeed run with the templating
applied.

Signed-off-by: Alexander Jung <alex@unikraft.io>