schema,unikraft/component: Propagate project workdir over core source
This commit changes the internal schema parser and the internal
`ComponentConfig` attribute `coreSource` to `workdir`.  Instead of
passing the Unikraft core's source code, pass to each component the
working directory.  Since the path of the core is a well-known, defined
through the exposed API `unikraft.PlaceComponent`, the core's source can
be determined.  Additionally and therefore, each component's source code
can also be subsequently determined.

This commit makes necessary changes to the schema loader to be able to
propgate this information.

Signed-off-by: Alexander Jung <alex@unikraft.io>