schema,unikraft/component: Embed `ComponentType` in `ComponentConfig`
This commit introduces a new option, `WithType`, for `ComponentConfig`
which embeds `unikraft.ComponentType` as an internal attribute.  Whilst
the `Component` interface does mandate an implementation method of
`Type() unikraft.ComponentType`, the interface `ComponentConfig` is also
an "extension" of each component and therefore the methods of
`ComponentConfig` extend 1-to-1 to any component, e.g.
`ApplicationConfig`.  Therefore, to contextualize these extended
methods, we pass in the type.

Signed-off-by: Alexander Jung <alex@unikraft.io>