exec: Introduce `exec` package
The top-level `exec` package is a process executor and manager which
provides an abstraction and mechanisms necessary to oversee processes
which will be executed by KraftKit.  The initial implementation provides
traditional uerspace process managemenet via `os.Exec`.  In the future,
injecting a build environment as an option (see #49) should be possible
with this same package.

When preparing a process to be executed, a new executable can be
generated which pre-fill CLI options through the interface structure tag
`flag`.  This is parsed via reflection, such that:

```go
type MyStruct struct {
    Attribute string `flag:"--atribute"`
}
```

Will ultimately pass the value of `Attribute` as a CLI option
`--attribute`.  This makes invoking a program more idiomatic and without
having to write extensive code for each unique program.  Simply passing
by value the structure to `NewExecutable` will generate the necessary
command line to be executed.

Signed-off-by: Alexander Jung <alex@unikraft.io>