make: Introduce `make` package
The top-level `make` package is an abstraction over GNU Make which has
provided via the package as a programmable API where each CLI option can
be evaluated via the `exec` package.  The package itself offers many
options including the ability to calculate the progress of an invocation
of make.  This is achieved by first calling `make` with the `-n` flag.

Signed-off-by: Alexander Jung <alex@unikraft.io>