unikraft/app: Introduce `Make` method to invoke Unikrat's build system
This commit introduces the method `Make` to `ApplicationConfig` which
ultimately invokes Unikraft's build system.  The necessary pieces are in
place to achieve this, including: knowing Unikraft's source directory
(ultimately the make CLI option `-C`) as well as the arguments to the
build system which are populated via `core.MakeArgs`.  This method
brings this together and allows programming the `make` package with
additional attributes.  Ultimately, well-known targets in Unikraft's
build system can use this as an underlying method and simply invoke a
specific target.

GitHub-Closes: #28
Signed-off-by: Alexander Jung <alex@unikraft.io>