log,internal/logger: Expose the main log writer as an interface method
This commit updates the `Logger` interface such that the `io.Writer`
which represents where the logger writes to is an attribute retrievable
externally so as to be referenced if it is re-configured (or cloned).
Additionally, we update the internal `Logger` such that the `io.Writer`
is now a private attribute.

Signed-off-by: Alexander Jung <alex@unikraft.io>