config: Migrate default config values to global constants
This commit migrates the value of the default value for a configuration
option to a globally accessible constant value.  The benefit of this
approach is that then without access to the configuration management
system (which may propagate values, or may not) the default value can
still be looked up, e.g.:

```go
import "kraftkit.sh/config"

config.DefaultLogLevel // "info"
```

These values can also be changed statically at compile-time with the
link-time flag `-X` for alternative usecases.

To accommodate this change, a globally accessible `Defaults()` method
returns a map of environmental variables to their default values.  The
environmental variable name is used as opposed to json or yaml values
since these are not embedded and simplifies lookup.

Signed-off-by: Alexander Jung <alex@unikraft.io>