machine/qemu{/qmp}: Introduce QEMU QMP Protobuf definitions
This commit introduces serialized definition of a limited set of method
calls and events from QEMU's "QAPI" known as QMP.  The definitions are
defined as Google Protobuf files and generated using the previously
introduced tool `protoc-gen-go-netconn`.

The QAPI protocol is relatively simple and is defined within QEMU's
source code in the `qapi/` folder.  The protocol is enabled by invoking
`-qmp` during the instantiation of the `qemu-system-*` binary.  In
server mode, the interface sends and receives JSON terminated by a
newline (`\n`).  The interface first expets a simple handshake which is
used to indicate the version of QEMU followed immediately by a
"capabilities" request: this is serialized in the "greeting" service.

By defining QAPI via Protobuf and with the associated tool, we are able
to maintain future additions to QAPI by providing the relevant service.
It is also easy to add existing interactions via Request-Response
format which is defined within `services.proto`.

This commit also includes the generated artifacts from the protoc
definitions which is done using `buf`[0].  In the future, these files
should be checked by the CI and also relevant generation of said files
should be added to the project's build system as a preparation step
before building the `kraft` binary.

Further on, it would be nice to automatically generate the QAPI
interface by directly parsing the source files from QEMU as they are
defiend in a relatively standard fashion.

[0]: https://buf.build/

Signed-off-by: Alexander Jung <alex@unikraft.io>