machine/qemu/qmp: Introduce `QMPEventMonitor` utility
This commit introduces a high-level and QEMU-version agnostic tool for
listening on a QMP interface.  The struct exposes an `Accept` method
which is used to listen on the interface.  The use of Go generics allows
us to plug-in any serialized QMP Protobuf which is generated to Go.

The event monitor does not yet support context cancellation.

Signed-off-by: Alexander Jung <alex@unikraft.io>