machine: Introduce `MachineStore` to save state and config of machines
This commit introduces a basic primitive to the "KraftKit Machine
Manager" where we are able to save information about the standard
`MachineConfig` which is used across all defined VMMs as well as the
driver-specific configuration (using the `any` generic).  The store uses
Badger[0] to facilite local key-value store of this information.  Each
method is concurrency safe.

Signed-off-by: Alexander Jung <alex@unikraft.io>