machine/qemu: Introduce global structure for holding QEMU config
This commit introduces the `QemuConfig` struct which is a serialization
of currently-supported QEMU options.  Each attribute is annotated with a
`flag` Go tag which is parsable by `kraftkit.sh/exec` and can be later
used during the instantiation of a QEMU instance.  Each option also has
a corresponding option which is used (and can be in the future enhanced)
to validate the contents of the value for the individual flag
pre-emptively.

Signed-off-by: Alexander Jung <alex@unikraft.io>