machine/qemu: Introduce the QEMU driver implementation
This commit implements `kraftkit.sh/machine/driver.Driver` for QEMU.

The implementation uses the serialization of necessary (and for now
minimal) arguments necessary for running Unikraft unikernels.  Each
supported flag appears within the `QemuConfig` and is ultimately
serialized (thanks to `fmt.Stringer`) into corresponding and correct
arguments by the `kraftkit.sh/exec` package.  The result will
instantiate a QEMU instance for the set of `machine.MachineOption`s.

The interaction with QEMU is performed via it's exposed QAPI protocol
(QMP).  When the instance is launched, it is always launched in "paused"
mode such that the driver's `Start()` method can later be used to
actually start the machine.

This implementation does not yet support proper disk or networking
support though this interaction should now be straightforward given the
infrastructure available: access to ad-hoc attach devices via QMP; and,
serialization of arguments via `exec`.

Signed-off-by: Alexander Jung <alex@unikraft.io>