manifest: Apply options correctly to the Manifest
This commit re-applies the options to the Manifest.  Currently it is
applied both before and after the remote is probed for versions and
channels.  The reason for this is to access the logger via the
`WithLogger` option.  In reality, this option should be deprecated in
favour of a globalized contextual logger which will remove every
instance of `WithLogger`.  See issue #74.  Once this issue is fixed, the
application of options should be removed before probbing.

Signed-off-by: Alexander Jung <alex@unikraft.io>