manifest,packmanager: Ability to search catalog by source and ephemerally
This commit introduces two new options to the `packmanager.CatalogQuery`
request which makes it possible to query the associated package manager
by its origin source as well as to use a just-in-time search by not
using any local cache.

The option `Source` is useful for querying the package manager if the
original source is known but not the package.  This means that the
abstractions associated with decompiling information such that it fits
into the previous search fields (`Name`, `Types`, etc.) do not need to
occur before the API.

The option `NoCache` forces the package manager to re-query internally
such that the package manager's `Update()` method does not need to
be called before invoking `Catalog()`.  This is also useful for any
just-in-time queries, e.g. for ephemeral builds which may not have
existing manifests to query.

Signed-off-by: Alexander Jung <alex@unikraft.io>