tui/paraprogress: Fix output and return the error from a failed progress
This commit handles a concurrency issue which occassionally occurs with
the TUI which would result in displaying the output of the console being
incorrectly displayed.  In addition to this, this commit makes the error
message from `Start()` to return the first occuring error from any
progress task.  This API therefore allows us to prevent any additional
execution which may occur after invoking the paraprogress package.

Signed-off-by: Alexander Jung <alex@unikraft.io>