manifest: Introduce mandatory `PullPackage` to `Provider`
This commit introduces a new interface method to `Provider` which
abstracts the process of retrieving the resource via the new method
`PullPackage`.  To accomplish this change, the provider is embedded
within the `ManifestPackage` and each provider is updated to facilitate
the retrieval.  In the current case, only the git and GitHub providers
are updated.  Additional information is also passed to the retrieval
process which now includes `pack.PackageOptions` as well as
`pack.PullPackageOptions`.  This is useful for contextualizing the pull
itself.

Signed-off-by: Alexander Jung <alex@unikraft.io>