manifest: Introduce `DirectoryProvider`
This commit introduces a new provider based on specifiying a path to an
existing directory on disk which represents a raw component.  This is
useful for local development of a component whilst not having to use
either git itself.  Assumptions are made which are provided to the
`Manifest` which is that there is only one channel, entitled `default`,
and that there are no versions.

Signed-off-by: Alexander Jung <alex@unikraft.io>