kconfig: Fix parsing Unikraft `Config.uk` files with no `menuend`
Unikraft's configuration file for microlibraries use a `Config.uk` which
expose configuration information, options, etc.  The prominant syntax of
these files do not necessarily end any `menuconfig` identifiers with a
closing `endmenu` identifier.  This is technically not strictly
necessarily based on the KConfig specification, however, the current
implementation uses this as a signifier to parse elements in a stack.
To accommodate for this, this commit adjusts how the stack is popped and
pushed such that we can ignore this trait entirely rather than error-ing
out.

Signed-off-by: Alexander Jung <alex@unikraft.io>