unikraft: Introduce mandatory `KConfigMenu` method to `Component`s
This commit returns the valid `KConfigMenu` which is exposed by a
microlibrary such that its whole configuration can be viewed.  For now,
the platform's menu is not returned as this can be both an external
microlibrary as well as from Unikraft's core.  To be able to explore the
core, both the core's version as well as the path to the core need to be
available.

Signed-off-by: Alexander Jung <alex@unikraft.io>