pack: Turn `Kernel` and `InitrdConfig` into first-class citizens
This commit removes the kernel and initrd from the metadata attribute
and places them as top-level attributes of `PacakgeOptions`.  Whilst not
all packages may necessarily have these particular files (for example,
in the future libraries are packages as well), this allows us to use the
package in a multi-facetted way where the `Type` is the switch.

Signed-off-by: Alexander Jung <alex@unikraft.io>