cmd/kraft: Propagate additional initializers to all subcommands
The initializer listed in `cmd/kraft/pkg/pkg.go` is required many
subcommands which rely on package management and exists as an artifact
from when each core subcommand was built as an inidividual binary.  In
this commit, we migrate this initializer to `cmd/kraft/kraft.go` which
then allows us to use it within every subcoommand.

Signed-off-by: Alexander Jung <alex@unikraft.io>